changed data export handling
changed graph-format CLI arg to be a list, in order to specify multiple
export types in a single go

made yaml one of the export types, rather than handling it as a separate
CLI argument

effectively disabled gephi/gexf export - the GEXF format doesn't support
complex attributes, so can't handle the data attached to the networks
that are handled with this script.  This may be a problem w/the networkx
GEXF export functions or it may be a problem w/the format itself.  For now, if
gephi is specified on the CLI as an graph-format this function will simply
log a message that the export format isn't currently supported.

In the future, need to fix this by "flattening" complex data types attached
to nodes, edges, the graph, etc.