collect security group rules in network at graph level
previously the security group rules were held in a separate dict.  That
dict is still used for all related functions - e.g. rule checks, etc.
but now there's a function get_sec_group_rules_by_vpc() that extracts
the same info as held in that "free floating" dictionary but stores it
in the networkx.Graph.graph dictionary, per VPC. At the moment this is only
for useful when a human examines the net dump file.  Eventually this will
be used for all other functions (i.e. the "free floating rule dict" will go
away).

The next step will be to run rule checks against this dict, rather than the
"free floating dict" because the latter may NOT have all the security groups
or rules in it.  This is b/c the current state trys to associate rules with
subnets, which is done indirectly by gathering sec-groups directly from
instances.  This may not result in all the sec-groups assocaited with the
instance b/c the API doesn't include sec-groups associated with anything
other than the instances primary network interface (i.e. sec-groups are
really associated with network interfaces not instances)